Reference

QFiND.AAAfittedSD

QFiND.BrownianSD

QFiND.DrudeSD

QFiND.PowerLawExpSD

QFiND.TannorMeyerSD

QFiND.balanced_truncation

QFiND.balanced_truncation

QFiND.bsdo_discr

QFiND.evaluate_error

QFiND.id_freq

QFiND.id_freq

QFiND.id_freq

QFiND.id_freq

QFiND.id_time

QFiND.nnls_weight

QFiND.orthpoly_discretization

Methods

QFiND.AAAfittedSD — Method

(specdens::AAAfittedSD)(ω::Float64; scale::Float64=1.0) -> Float64

Compute the spectral density for the Brownian oscillator model.

QFiND.BrownianSD — Method

(specdens::BrownianSD)(ω::Float64; scale::Float64=1.0) -> Float64

Compute the spectral density for the Brownian oscillator model.

QFiND.DrudeSD — Method

(specdens::DrudeSD)(ω::Float64; scale::Float64=1.0) -> Float64

Compute the spectral density for the Drude-Lorentz model.

QFiND.PowerLawExpSD — Method

(specdens::PowerLawExpSD)(ω::Float64; scale::Float64=1.0) -> Float64

Compute the spectral density for the Power-law with exponential cutoff model.

QFiND.TannorMeyerSD — Method

(specdens::TannorMeyerSD)(ω::Float64; scale::Float64=1.0) -> Float64

Compute the spectral density for the Tannor-Meyer model.

QFiND.balanced_truncation — Method

balanced_truncation(a::AbstractVector{<:Number}, c::AbstractVector{<:Number}, eps::Float64) :: Exponentials

Given the exponents a and coefficients c of a sum of exponentials, compute a new sum of exponentials with a reduced number of terms for a given tolerance eps. The initial exponents must have positive real part.

QFiND.balanced_truncation — Method

balanced_truncation(a::AbstractVector{<:Number}, c::AbstractVector{<:Number}, M::Int) :: Exponentials

Given the exponents a and coefficients c of a sum of exponentials, compute a new sum of exponentials with a given number of terms M. The initial exponents must have positive real part.

QFiND.bsdo_discr — Method

bsdo(sbeta, ω::AbstractVector{Float64}, M_sp::Int)

Compute the discretization of the QNSD using the BSDO method.

QFiND.evaluate_error — Method

Calculate the error between the approximated and exact correlation functions.

QFiND.id_freq — Method

id_freq_rank(f::AbstractMatrix{ComplexF64}, frank::Int, rnd::Bool)

Perform interpolative decomposition on matrix f with a specified rank frank.

QFiND.id_freq — Method

id_freq_eps(f::AbstractMatrix{Float64}, eps::Real, rnd::Bool)

Perform interpolative decomposition on matrix f with error tolerance eps.

QFiND.id_freq — Method

id_freq_rank(f::AbstractMatrix{Float64}, frank::Int, rnd::Bool)

Perform interpolative decomposition on matrix f with a specified rank frank.

QFiND.id_freq — Method

id_freq_eps(f::AbstractMatrix{Float64}, eps::Real, rnd::Bool)

Perform interpolative decomposition on matrix f with error tolerance eps.

QFiND.id_time — Method

id_time(f::AbstractMatrix{ComplexF64}, frank::Int, rnd::Bool)

Perform interpolative decomposition on matrix f with a specified rank frank.

QFiND.nnls_weight — Method

nnls_weight(t, B)

Estimate the coefficients (amplitudes) by solving a nonnegative least‐squares (NNLS) problem.

QFiND.orthpoly_discretization — Method

orthpoly_discretization(wj, N)

Compute the discretization of orthogonal polynomials.